+++ src/palo-alto-networks/azext_palo_alto_networks/aaz/latest/palo_alto/cloudngfw/local_rulestack/fqdnlist/_wait.py
@@ -15,7 +15,7 @@ from azure.cli.core.aaz import *
     "palo-alto cloudngfw local-rulestack fqdnlist wait",
 )
 class Wait(AAZWaitCommand):
-    """Place the CLI in a waiting state until a condition is met.
+    """Wait until a specified condition is met for the Palo Alto Networks local rulestack FQDN list operation to complete.
     """
 
     _aaz_info = {